feat(server_session): clients can no longer re-use message ids
There is a test and it passes.
